requirements.txt
certifi==2020.6.20
chardet==3.0.4
idna==2.10
numpy==1.16.6
Pillow==6.2.2
pkg-resources==0.0.0
requests==2.24.0
scipy==1.2.3
urllib3==1.25.11
